+++ tests/src/IntegrationTest/MapAttributeTest.php
@@ -14,6 +14,8 @@ declare(strict_types=1);
 namespace Rekalogika\Mapper\Tests\IntegrationTest;
 
 use Rekalogika\Mapper\Tests\Common\FrameworkTestCase;
+use Rekalogika\Mapper\Tests\Fixtures\MapAttribute\ObjectExtendingSomeObjectDto;
+use Rekalogika\Mapper\Tests\Fixtures\MapAttribute\ObjectOverridingSomeObjectDto;
 use Rekalogika\Mapper\Tests\Fixtures\MapAttribute\SomeObject;
 use Rekalogika\Mapper\Tests\Fixtures\MapAttribute\SomeObjectDto;
 
@@ -28,4 +30,25 @@ class MapAttributeTest extends FrameworkTestCase
         $this->assertEquals('propertyB', $target->getTargetPropertyB());
         $this->assertEquals('propertyC', $target->getTargetPropertyC());
     }
+
+    public function testMapAttributeOnSubclass(): void
+    {
+        $source = new SomeObject();
+        $target = $this->mapper->map($source, ObjectExtendingSomeObjectDto::class);
+
+        $this->assertEquals('propertyA', $target->targetPropertyA);
+        $this->assertEquals('propertyB', $target->getTargetPropertyB());
+        $this->assertEquals('propertyC', $target->getTargetPropertyC());
+    }
+
+    public function testMapAttributeOnOverridingSubclass(): void
+    {
+        $source = new SomeObject();
+        $target = $this->mapper->map($source, ObjectOverridingSomeObjectDto::class);
+
+        $this->assertEquals('propertyB', $target->targetPropertyA);
+        $this->assertEquals('propertyC', $target->getTargetPropertyB());
+        $this->assertEquals('propertyA', $target->getTargetPropertyC());
+    }
+
 }